// src/interfaces/commodity.interface.ts
// Define the interface for the Commodity
export interface Commodity {
    commodity_id: number;
  commodity_name: string;
  commodity_type_id: number;
  image_id: number;
  status: number;
  created_on: Date;
  created_by: number;
  updated_on: Date;
  updated_by: number;
  commodity_code: string;
  }
  
  
  